raspberrypi/requirements.txt
pyserial
rx